fix(TestSiteDomain): Move test over to TestSite
Rename callback is faked correctly in TestSite
