+++ press/press/doctype/site/test_site.py
@@ -302,3 +302,20 @@ class TestSite(unittest.TestCase):
 		domains = site.get_config_value_for_key("domains")
 		self.assertNotIn(domain, domains)
 		self.assertIn(domain_2, domains)
+
+	def test_site_rename_doesnt_update_host_name_for_custom_domain(self):
+		"""Ensure site configuration isn't updated after rename when custom domain is host_name."""
+		from press.press.doctype.site_domain.test_site_domain import create_test_site_domain
+
+		site = create_test_site("old-name")
+		site_domain1 = create_test_site_domain(site.name, "sitedomain1.com")
+		site.set_host_name(site_domain1.name)
+		new_name = "new-name.fc.dev"
+		site.rename(new_name)
+
+		rename_job = self._fake_succeed_rename_jobs()
+		process_rename_site_job_update(rename_job)
+		site = frappe.get_doc("Site", new_name)
+		if site.configuration[0].key == "host_name":
+			config_host = site.configuration[0].value
+		self.assertEqual(config_host, f"https://{site_domain1.name}")
